requirements.txt
asgiref==3.7.2
colorama==0.4.6
coverage==7.3.1
Django==4.2.5
django-cleanup==8.0.0
django-dotenv==1.4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
exceptiongroup==1.1.3
iniconfig==2.0.0
packaging==23.1
Pillow==10.0.1
pluggy==1.3.0
PyJWT==2.8.0
pytz==2023.3.post1
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.8.0
tzdata==2023.3
